fix: share layout between /sites/** and /settings
