+++ frontend/app/[lang]/(main)/sites/add/page.tsx
@@ -1,5 +1,4 @@
 import TitleBar from '@/app/components/onboarding/TitleBar';
-import { Box } from '@mui/material';
 import { Dict, getDictonaryWithDefault } from '@/app/[lang]/dictionaries';
 import { ArrowBack } from '@mui/icons-material';
 import SiteForm from '@/app/components/forms/SiteForm';
@@ -11,16 +10,7 @@ const AddSitePage = async ({
 }) => {
   const dict: Dict = getDictonaryWithDefault(lang);
   return (
-    <Box
-      sx={{
-        height: '100%',
-        width: '100%',
-        display: 'flex',
-        flexDirection: 'column',
-        justifyContent: 'space-between',
-        padding: '32px 32px 40px 32px',
-      }}
-    >
+    <>
       <TitleBar
         dict={dict}
         icon={<ArrowBack fontSize="small" />}
@@ -28,7 +18,7 @@ const AddSitePage = async ({
         href={`/${lang}/sites`}
       />
       <SiteForm dict={dict} redirectPath={`/${lang}/sites`} />
-    </Box>
+    </>
   );
 };
